pyproject.toml
[project]
name = "new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "av>=14.4.0",
    "groq>=0.31.1",
    "joblib>=1.5.2",
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.7",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "streamlit>=1.49.1",
    "streamlit-webrtc>=0.63.4",
]
